feat(editor): add page dragging area widget extension (#12045)
Closes: BS-3364
