+++ blocksuite/affine/all/package.json
@@ -61,6 +61,7 @@
     "@blocksuite/affine-widget-frame-title": "workspace:*",
     "@blocksuite/affine-widget-keyboard-toolbar": "workspace:*",
     "@blocksuite/affine-widget-linked-doc": "workspace:*",
+    "@blocksuite/affine-widget-page-dragging-area": "workspace:*",
     "@blocksuite/affine-widget-remote-selection": "workspace:*",
     "@blocksuite/affine-widget-scroll-anchoring": "workspace:*",
     "@blocksuite/affine-widget-slash-menu": "workspace:*",
@@ -188,6 +189,8 @@
     "./widgets/keyboard-toolbar/view": "./src/widgets/keyboard-toolbar/view.ts",
     "./widgets/viewport-overlay": "./src/widgets/viewport-overlay/index.ts",
     "./widgets/viewport-overlay/view": "./src/widgets/viewport-overlay/view.ts",
+    "./widgets/page-dragging-area": "./src/widgets/page-dragging-area/index.ts",
+    "./widgets/page-dragging-area/view": "./src/widgets/page-dragging-area/view.ts",
     "./fragments/doc-title": "./src/fragments/doc-title.ts",
     "./fragments/frame-panel": "./src/fragments/frame-panel.ts",
     "./fragments/outline": "./src/fragments/outline.ts",
