+++ blocksuite/affine/all/src/widgets/page-dragging-area/index.ts
@@ -1,0 +1,1 @@
+export * from '@blocksuite/affine-widget-page-dragging-area';
