+++ blocksuite/affine/all/src/widgets/page-dragging-area/view.ts
@@ -1,0 +1,1 @@
+export * from '@blocksuite/affine-widget-page-dragging-area/view';
