--- blocksuite/affine/blocks/root/src/widgets/index.ts
@@ -1,1 +1,0 @@
-export { AffinePageDraggingAreaWidget } from './page-dragging-area/page-dragging-area.js';
