+++ blocksuite/affine/widgets/page-dragging-area/src/effects.ts
@@ -1,0 +1,11 @@
+import {
+  AFFINE_PAGE_DRAGGING_AREA_WIDGET,
+  AffinePageDraggingAreaWidget,
+} from './index';
+
+export function effects() {
+  customElements.define(
+    AFFINE_PAGE_DRAGGING_AREA_WIDGET,
+    AffinePageDraggingAreaWidget
+  );
+}
